# Conflicts:
#	.pre-commit-config.yaml
#	include/aie/Dialect/AIE/IR/AIEOps.td
#	lib/Dialect/AIEX/IR/AIEXDialect.cpp
#	python/iron/worker.py

+++ include/aie/Targets/AIETargetShared.h
@@ -34,7 +34,7 @@ std::string packetStr(int id, int type);
 void generateXAieDmaSetMultiDimAddr(llvm::raw_ostream &output, int ndims,
                                     llvm::ArrayRef<BDDimLayoutAttr> dims,
                                     int col, int row, int bdNum, int baseAddrA,
-                                    int offsetA, int lenA,
+                                    int64_t offsetA, uint32_t lenA,
                                     int elementWidthInBytes,
                                     const char *errorRet);
